fixed warning

+++ example/main.c
@@ -17,7 +17,7 @@ static void message_arrived(lwmqtt_client_t *c, lwmqtt_string_t *t, lwmqtt_messa
     exit(1);
   }
 
-  if (strncmp(payload, m->payload, m->payload_len) != 0) {
+  if (strncmp(payload, m->payload, (size_t)m->payload_len) != 0) {
     printf("payload does not match\n");
     exit(1);
   }
